python2Packages.faulthandler: 3.1 -> 3.2 and fix build
It is currently failing on master and 20.03. I spent some time reading the src
code but was not able to figure out why the env var activation is not working.
Since this is currently failing, and since it's dying alongside python 2 anyways
I propose we just disable the 1 failing test.

There's some more information inline in the comment if someone wants to fix this
or dig further.