nixos.xfce: use 'XFCE' as the session name
The desktop session name is used for the 'DesktopNames' field in the
generated session file (xfce.desktop). The standard desktop name for
Xfce is 'XFCE' (all in caps). The value for the 'XDG_CURRENT_DESKTOP'
environment variable is obtained from this field. When this variable
does not have the right value, some features do not work
correctly. For instance, menu items set to show only on Xfce are not
shown.