ghc-8.4.4.nix: Do not use git.haskell.org
which was deprecated in 2018 and is now gone for good. I guess many
won’t notice because the nix-cache kept the files around?

This fetches the patches from github. Unfortunately it seems that the
sha256 hashes (calculated with `nix-prefetch-url`) have been modified.

Alternatively, if someone has the old patches around they could arguably
be added to `tarballs.nixos.org` (it seems they are not there right
now).