Force "dev" output in case of static builds
Static builds propagate all buildInputs resulting in a
`nix-support/propagated-build-inputs` for nearly every package,
retaining references to the entire build-closure. To avoid this, one
needs to add "dev" outputs to the package.

This commit forces "dev" outputs in case of static builds. That is,
static builds will fail to build without a functioning "dev" output.

This will break packages that require the `propagated-build-inputs` file
in the main output. An example here are Python packages.

https://github.com/NixOS/nixpkgs/issues/83667