documentation.nixos.enable: Default to false
Generation of configuration.nix is hugely expensive. For example,
evaluation of the configuration of my laptop takes 3.97s realtime with
documentation.nixos.enable set to 'true', but 2.42s with it set to
'false'. In other words, documentation generation accounts for 39% of
evaluation time.

Given that the docs are readily available online (and in a more
convenient form than 'man configuration.nix'), it seems better to
disable them by default.