squashfs-tools-ng: 0.8 -> 0.9
Like all supported formats (I think), LZO is optional.  I don't see
any reason to bother with options for now, so I've just enabled it.  Plus,
there's a bug in this release that makes compilation fail if LZO is
disabled due to a missing include.  I've sent a fix[1] for that
upstream.

[1]: https://github.com/AgentD/squashfs-tools-ng/pull/41