nixos/zfs-replication.nix: expose port in config.
I found while attempting to use a nonstandard SSH port with this module
that the port hadn't been passed as an option in this module.  This
corrects that oversight and makes it configurable.

I've verified this with the following steps:

1. sudo nixos-rebuild -I .../nixpkgs switch
1. systemctl cat zfs-replication.service

The service did include the changed port and continues to execute.