nixos/networkmanager: make all plugins optional
This solves an issue with the current NetworkManager module where it's
impossible to use it on a system with `environment.noXlibs = true` due
to failures in some plugin builds.

Moreover, these are entirely optional and not crucial to having NM
working and I suspect most don't really need all, possibly any, of these
that were built-in.

I also attempted to clean up plugin-related side effects that weren't
within an "if cfg..." guard, again to minimize the default footprint of
the NM module.