stdenv: add _accumFlagsArray, for structured-attrs-compatible code
We'll use this shell function in a number of spots in setup.sh
that use a pattern like this (often with still more items in it):

    local flagsArray=(
        SHELL=$SHELL
        $makeFlags ${makeFlagsArray+"${makeFlagsArray[@]}"}
    )

which assumes `makeFlags` is a string to be split, and can
instead say:

    local flagsArray=(
        SHELL=$SHELL
    )
    _accumFlagsArray makeFlags makeFlagsArray

in order to behave correctly both with and without structured attrs.

In every case this pattern appears, the local variable is named
`flagsArray`.  So we keep this function a little simpler to read
by baking that name into its interface.

Similarly, in a lot of cases there are some input variables that
should be treated as arrays even in the old-API case; and every
such input variable has a name ending in `Array` and vice versa.
So we bake that convention into the interface too, in order to
save the complexity of specifying which to treat which way.

That all only makes sense as long as this function is for
stdenv's internal use.  If we find we want to use it more widely
(and take off the leading `_` from the name), it'll be better to
make the interface more explicit and just eat the complexity that
that adds.