stdenv: __structuredAttrs is now often usable!
Specifically, I believe that for the majority of simple packages,
enabling it will Just Work, with no further effort.

That's based on taking a random sample of simple packages, building
with and without, and carefully comparing the build logs.  (With a
tool to normalize out boring differences.)  In the sample, all built
successfully, and ~90% had no differences detectable in the build log.

For that sample, a "simple package" was defined as one where (a) the
expression directly calls `stdenv.mkDerivation`, not another
abstraction; and (b) the file is <=30 lines long, which puts it in
the smallest 30% or so of such packages.