pcsc-cyberjack: Fix build against libusb
Commit 59281f742a0ce8def71be3576129b4178ec579bb changed libusb so that
it no longer propagates libusb1. This in turn caused the pcsc-cyberjack
build to fail with an error like this:

  usbdev.c:68:3: error: #error "Neither HAL nor USB1 found!. Please install at least either of them."

I'm committing this directly to master, since I'm the maintainer of the
package and the build also is *already* broken. So the worst case that
could happen is that the build is still broken afterwards.

Signed-off-by: aszlig <aszlig@nix.build>