python3Packages.qiskit-terra: 0.12.0 -> 0.13.0
Also add in optional visualization inputs as default to ensure that
they are in the environment when used at runtime.
Some hacks to avoid packaging retworkx (uses a semi-complicated Rust ->
Python packaging system, will try to get that packaged in future).
For now, just disabling retworkx, will need to package it in future.