dell-command-configure: init at 4.2.0
As is, this matches how Dell Command Configure is distributed. If
openmanage/srvadmin were to be packaged for NixOS, it would be more
correct to pull srvadmin-hapi out and have it as a dependency, though I
haven't done any testing on whether dcc depends specifically on the
version it is distributed with. Latest version of srvadmin-hapi is 9.4.0

The deb and rpm packages additionally come with a systemd service to
preload some drivers, but when getting this to work, it was unnecessary
to get things to work. As this service works via a complex oneshot shell
script I opted to leave it out, but it might prove necessary for some
combinations of systems.

See:
DCC https://www.dell.com/support/article/us/en/19/sln311302/dell-command-configure
Documentation https://www.dell.com/support/home/us/en/dedhs1/product-support/product/command-configure-v4.2/docs
Openmanage http://linux.dell.com/repo/hardware/omsa.html