module/dell-command-configure: init
Creates the necessary configuration file to launch cctk. The cfg file is
mostly undocumented, and while cctk appears to run fine with just
`hapi.vardatapath` set, but this is (approximately, minus nix-related
changes) what is distributed with the package.

The program never actually wrote to those paths in my testing, and
appears to operate just fine with only that one variable set, but I did
not extensively test every single option, and certainly not across every
system that is supported by the software.