gst_all_1.gst-plugins-bad: disable zbar by default
gst-plugins-bad by default used to pull in gtk3 and qtbase and qtx11extras because of the default dependency on zbar.
As zbar is a rarely needed gstreamer plugin, this unnecessarily
increased the closure size.
I am only aware of gnome-keysign actually using the zbar plugin, so that
uses a zbar-enabled gst-plugins-bad.

closes #84845