qimgv: enable video support
I actually went back and `strace`'d the binary when it complained about
being unable to load the library it uses for playing videos. Turns out,
it wasn't finding the library because it wasn't in any of its library
paths. I added the lib path to `LD_LIBRARY_PATH` so it can find the
library it uses to play videos, and now things are peachy.

There is a (seemingly innocuous) error that gets displayed from Exiv2
being unable to determine its image type. Since it's actually a video, I
think it's fine. Another issue that pops up in the output is missing
`libcuda.so.1`, but that doesn't seem to affect functionality, and using
`addOpenGLRunpath` on both the binary and libraries didn't silence it.

Also did a little formatting.