nixos/installer: Make nixos-rebuild test and dry-activate not create a symlink in the current directory
The section for `test`, `build`, `dry-build`, and `dry-activate` has been
split into two separate cases, one for `build` and one for the rest. The
handling for `build` is identical to how it was before, but the others use
--no-out-link to not create a result symlink in the current directory.

Note that this actually doesn't matter for `dry-build`. I've grouped `dry-build`
with `test` and `dry-activate` since it shouldn't create a ./result symlink
(which it doesn't do anyway)